requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
PyYAML==6.0.2
SQLAlchemy==2.0.35
sqlmodel==0.0.22
pdfplumber==0.11.4
pikepdf==9.2.1
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httpx==0.27.2
structlog==24.4.0
apscheduler==3.10.4
python-multipart==0.0.12
